norminetted and refactored, working version, can even compile and launch minishell in minishell
