+++ Makefile
@@ -30,6 +30,7 @@ MINISHELL_SRC= 							\
 	src/output/redirection.c			\
 	src/parser/parser_utils.c			\
 	src/parser/command_table.c			\
+	src/parser/command_table_utils.c	\
 	src/parser/dispatcher.c				\
 	src/parser/get_executable_path.c	\
 	src/commands/exit_command.c			\
