+++ src/output/handle_pipes.c
@@ -7,26 +7,12 @@
 #include <parser/dispatcher.h>
 #include <commands/echo_utils.h>
 
-/*
-	assuming the full unknown command is passed with full checked executable
-	path as this is already checked when parsing to see if given
-	command is valid ex:
-		commands->arg->start = "...[c]at -e main..."
-		command->arg->end = "...cat -e mai[n]..."
-		command->exe_path = "/bin/cat"
-*/
-
 void	create_table(t_command commands[], char *arg, char *path)
 {
 	commands->arg.start = arg;
 	commands->arg_len = strlen(arg);
 	commands->exe_path = path;
 }
-
-/*
-	If command_code == valid_different split arguments
-	// execute_commands(&commands[i], env) to be place in dispatcher
-*/
 
 int	run_multi_processes(const char *env[],
 	t_command commands[], int num_of_processes)
@@ -43,7 +29,7 @@ int	run_multi_processes(const char *env[],
 		if (pid == CHILD_PROCESS)
 		{
 			redirect_in_and_output(&pipes, i, num_of_processes);
-			dispatch_commands(&commands[i], env);
+			dispatch_command(&commands[i], env);
 		}
 		if (i != FIRST_PROCESS)
 			close(pipes.previous[READ_FD]);
