+++ src/parser/command_table.h
@@ -6,5 +6,7 @@
 t_command_code	get_command_code(const char **input, t_command *command);
 t_bool			is_between_quotes(const char *input, int reserved_char_index);
 int				get_arg_len(const char *start);
+int				get_cmd_len(const char *input);
+t_bool			is_system_command(const char *input, t_command *command);
 
 #endif
