+++ src/parser/dispatcher.c
@@ -40,7 +40,7 @@ t_exit_code	empty_command(t_command command, t_output_stdout write_to_stdout)
 	return (SUCCESS);
 }
 
-t_exit_code	dispatch_commands(const t_command *command, const char *env[])
+t_exit_code	dispatch_command(const t_command *command, const char *env[])
 {
 	static const t_command_function		functions[LAST] = {
 															empty_command,
@@ -49,18 +49,8 @@ t_exit_code	dispatch_commands(const t_command *command, const char *env[])
 															pwd_command,
 															unknown_command,
 															};
+
 	if (command->code == SYSTEM)
 		execute_commands(command, env);
 	exit(functions[command->code](*command, write_to_stdout));
 }
-
-/*
-
-t_command_function commands[LAST] = {
-										empty_command,
-										echo_command,
-										exit_command,
-										pwd_command,
-									}
-
-*/
