+++ src/parser/dispatcher.h
@@ -6,6 +6,6 @@
 typedef t_exit_code(*t_command_function)(t_command, t_output_stdout);
 
 t_exit_code	unknown_command(t_command command, t_output_stdout output);
-t_exit_code	dispatch_commands(const t_command *command_table, const char *env[]);
+t_exit_code	dispatch_command(const t_command *command, const char *env[]);
 
 #endif
